Move connected user count to Authentication.
